refactor: move get_specs_paths
